Create 3DPoc.Viewmodel project Move some of the base viewmodels and helpers
