+++ .gitignore
@@ -7,3 +7,4 @@
 3DPoC/POC3D/POC3D.Model/obj/
 3DPoC/POC3D/POC3D.Model.Tests/bin/
 3DPoC/POC3D/POC3D.Model.Tests/obj/
+3DPoC/POC3D/POC3D.ViewModel/obj/
