+++ 3DPoC/POC3D/POC3D.ViewModel/Base/Observable.cs
