+++ 3DPoC/POC3D/POC3D.ViewModel/Geometry/GraphicsHelper.cs
@@ -1,7 +1,7 @@
 ﻿using System.Windows.Media;
 using System.Windows.Media.Media3D;
 
-namespace POC3D.Helpers
+namespace POC3D.ViewModel.Geometry
 {
     public static class GraphicsHelper
     {
