+++ 3DPoC/POC3D/POC3D.sln
@@ -11,6 +11,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{727E8E2D
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "POC3D.Model.Tests", "POC3D.Model.Tests\POC3D.Model.Tests.csproj", "{F038248C-EA0C-465A-8D8B-F84A012BF864}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "POC3D.ViewModel", "POC3D.ViewModel\POC3D.ViewModel.csproj", "{011C1C91-287C-4502-9647-91E844843630}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -29,6 +31,10 @@ Global
 		{F038248C-EA0C-465A-8D8B-F84A012BF864}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F038248C-EA0C-465A-8D8B-F84A012BF864}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F038248C-EA0C-465A-8D8B-F84A012BF864}.Release|Any CPU.Build.0 = Release|Any CPU
+		{011C1C91-287C-4502-9647-91E844843630}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{011C1C91-287C-4502-9647-91E844843630}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{011C1C91-287C-4502-9647-91E844843630}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{011C1C91-287C-4502-9647-91E844843630}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
