+++ 3DPoC/POC3D/POC3D/Controls/VisualAreaControl.xaml.cs
@@ -5,8 +5,8 @@ using System.Windows.Controls;
 using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Media3D;
-using POC3D.Helpers;
 using POC3D.ViewModel;
+using POC3D.ViewModel.Geometry;
 
 namespace POC3D.Controls
 {
