+++ 3DPoC/POC3D/POC3D/ViewModel/ElementViewModel.cs
@@ -4,6 +4,7 @@ using System.Windows.Media.Media3D;
 using POC3D.Helpers;
 using POC3D.Model;
 using POC3D.Model.Calculations;
+using POC3D.ViewModel.Geometry;
 
 namespace POC3D.ViewModel
 {
