+++ 3DPoC/POC3D/POC3D/ViewModel/ResultElementViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.Windows.Media;
 using System.Windows.Media.Media3D;
-using POC3D.Helpers;
+using POC3D.ViewModel.Geometry;
 
 namespace POC3D.ViewModel
 {
